feat(site/AgentsPage): merge Limits + Usage into unified Spend page
Merge the agent settings Limits and Usage pages into a single Spend page
with collapsible card sections for default limits, group limits, and
per-user spend. Add CollapsibleSection component, URL-based date range
picker on Analytics page, redirect old routes, and standardize layout
spacing across all agent settings pages.
