+++ site/src/pages/AgentsPage/AgentAnalyticsPageView.tsx
@@ -1,15 +1,21 @@
-import { BarChart3Icon } from "lucide-react";
 import type { FC } from "react";
 import type { ChatCostSummary } from "#/api/typesGenerated";
+import {
+	DateRangePicker,
+	type DateRangeValue,
+} from "#/components/DateRangePicker/DateRangePicker";
 import { ChatCostSummaryView } from "./components/ChatCostSummaryView";
 import { SectionHeader } from "./components/SectionHeader";
+import { toInclusiveDateRange } from "./utils/dateRange";
 
 interface AgentAnalyticsPageViewProps {
 	summary: ChatCostSummary | undefined;
 	isLoading: boolean;
 	error: unknown;
 	onRetry: () => void;
-	rangeLabel: string;
+	dateRange: DateRangeValue;
+	onDateRangeChange: (value: DateRangeValue) => void;
+	hasExplicitDateRange: boolean;
 }
 
 export const AgentAnalyticsPageView: FC<AgentAnalyticsPageViewProps> = ({
@@ -17,8 +23,15 @@ export const AgentAnalyticsPageView: FC<AgentAnalyticsPageViewProps> = ({
 	isLoading,
 	error,
 	onRetry,
-	rangeLabel,
+	dateRange,
+	onDateRangeChange,
+	hasExplicitDateRange,
 }) => {
+	const displayDateRange = toInclusiveDateRange(
+		dateRange,
+		hasExplicitDateRange,
+	);
+
 	return (
 		<div className="flex flex-col p-4 pt-8">
 			<div className="mx-auto w-full max-w-3xl">
@@ -26,10 +39,10 @@ export const AgentAnalyticsPageView: FC<AgentAnalyticsPageViewProps> = ({
 					label="Analytics"
 					description="Review your personal Coder Agents usage and cost breakdowns."
 					action={
-						<div className="flex items-center gap-2 text-xs text-content-secondary">
-							<BarChart3Icon className="h-4 w-4" />
-							<span>{rangeLabel}</span>
-						</div>
+						<DateRangePicker
+							value={displayDateRange}
+							onChange={onDateRangeChange}
+						/>
 					}
 				/>
