+++ site/src/pages/AgentsPage/AgentSettingsSpendPage.tsx
@@ -1,14 +1,30 @@
 import dayjs from "dayjs";
 import { type FC, useState } from "react";
-import { keepPreviousData, useQuery } from "react-query";
+import {
+	keepPreviousData,
+	useMutation,
+	useQuery,
+	useQueryClient,
+} from "react-query";
 import { useSearchParams } from "react-router";
-import { chatCostSummary, chatCostUsers } from "#/api/queries/chats";
+import {
+	chatCostSummary,
+	chatCostUsers,
+	chatUsageLimitConfig,
+	deleteChatUsageLimitGroupOverride,
+	deleteChatUsageLimitOverride,
+	updateChatUsageLimitConfig,
+	upsertChatUsageLimitGroupOverride,
+	upsertChatUsageLimitOverride,
+} from "#/api/queries/chats";
+import { groups } from "#/api/queries/groups";
 import { user } from "#/api/queries/users";
+import type { ChatCostUserRollup } from "#/api/typesGenerated";
 import type { DateRangeValue } from "#/components/DateRangePicker/DateRangePicker";
 import { useDebouncedValue } from "#/hooks/debounce";
 import { useAuthenticated } from "#/hooks/useAuthenticated";
 import { RequirePermission } from "#/modules/permissions/RequirePermission";
-import { AgentSettingsUsagePageView } from "./AgentSettingsUsagePageView";
+import { AgentSettingsSpendPageView } from "./AgentSettingsSpendPageView";
 
 const pageSize = 10;
 
@@ -23,19 +39,44 @@ const getDefaultUsageDateRange = (now?: dayjs.Dayjs): DateRangeValue => {
 	};
 };
 
-interface AgentSettingsUsagePageProps {
+interface AgentSettingsSpendPageProps {
 	/** Override the current time for date range calculation. Used for
 	 *  deterministic Storybook snapshots. */
 	now?: dayjs.Dayjs;
 }
 
-const AgentSettingsUsagePage: FC<AgentSettingsUsagePageProps> = ({ now }) => {
+const AgentSettingsSpendPage: FC<AgentSettingsSpendPageProps> = ({ now }) => {
 	const { permissions } = useAuthenticated();
+	const queryClient = useQueryClient();
+
+	// --------------- Limits queries & mutations ---------------
+
+	const configQuery = useQuery(chatUsageLimitConfig());
+	const groupsQuery = useQuery(groups());
+
+	const updateConfigMutation = useMutation(
+		updateChatUsageLimitConfig(queryClient),
+	);
+	const upsertOverrideMutation = useMutation(
+		upsertChatUsageLimitOverride(queryClient),
+	);
+	const deleteOverrideMutation = useMutation(
+		deleteChatUsageLimitOverride(queryClient),
+	);
+	const upsertGroupOverrideMutation = useMutation(
+		upsertChatUsageLimitGroupOverride(queryClient),
+	);
+	const deleteGroupOverrideMutation = useMutation(
+		deleteChatUsageLimitGroupOverride(queryClient),
+	);
+
+	// --------------- Usage state & queries ---------------
 
 	const [searchParams, setSearchParams] = useSearchParams();
 	const [searchFilter, setSearchFilter] = useState("");
 	const debouncedSearch = useDebouncedValue(searchFilter, 300);
 	const [page, setPage] = useState(1);
+
 	const startDateParam =
 		searchParams.get(usageStartDateSearchParam)?.trim() ?? "";
 	const endDateParam = searchParams.get(usageEndDateSearchParam)?.trim() ?? "";
@@ -100,7 +141,51 @@ const AgentSettingsUsagePage: FC<AgentSettingsUsagePageProps> = ({ now }) => {
 
 	return (
 		<RequirePermission isFeatureVisible={permissions.editDeploymentConfig}>
-			<AgentSettingsUsagePageView
+			<AgentSettingsSpendPageView
+				// Limits config
+				configData={configQuery.data}
+				isLoadingConfig={configQuery.isLoading}
+				configError={configQuery.isError ? configQuery.error : null}
+				refetchConfig={() => void configQuery.refetch()}
+				groupsData={groupsQuery.data}
+				isLoadingGroups={groupsQuery.isLoading}
+				groupsError={groupsQuery.isError ? groupsQuery.error : null}
+				onUpdateConfig={updateConfigMutation.mutate}
+				isUpdatingConfig={updateConfigMutation.isPending}
+				updateConfigError={
+					updateConfigMutation.isError ? updateConfigMutation.error : null
+				}
+				isUpdateConfigSuccess={updateConfigMutation.isSuccess}
+				resetUpdateConfig={updateConfigMutation.reset}
+				onUpsertOverride={({ userID, req, onSuccess }) =>
+					upsertOverrideMutation.mutate({ userID, req }, { onSuccess })
+				}
+				isUpsertingOverride={upsertOverrideMutation.isPending}
+				upsertOverrideError={
+					upsertOverrideMutation.isError ? upsertOverrideMutation.error : null
+				}
+				onDeleteOverride={deleteOverrideMutation.mutate}
+				isDeletingOverride={deleteOverrideMutation.isPending}
+				deleteOverrideError={
+					deleteOverrideMutation.isError ? deleteOverrideMutation.error : null
+				}
+				onUpsertGroupOverride={({ groupID, req, onSuccess }) =>
+					upsertGroupOverrideMutation.mutate({ groupID, req }, { onSuccess })
+				}
+				isUpsertingGroupOverride={upsertGroupOverrideMutation.isPending}
+				upsertGroupOverrideError={
+					upsertGroupOverrideMutation.isError
+						? upsertGroupOverrideMutation.error
+						: null
+				}
+				onDeleteGroupOverride={deleteGroupOverrideMutation.mutate}
+				isDeletingGroupOverride={deleteGroupOverrideMutation.isPending}
+				deleteGroupOverrideError={
+					deleteGroupOverrideMutation.isError
+						? deleteGroupOverrideMutation.error
+						: null
+				}
+				// Usage data
 				dateRange={dateRange}
 				hasExplicitDateRange={hasExplicitDateRange}
 				onDateRangeChange={onDateRangeChange}
@@ -128,7 +213,7 @@ const AgentSettingsUsagePage: FC<AgentSettingsUsagePageProps> = ({ now }) => {
 						return next;
 					});
 				}}
-				onSelectUser={(u) => {
+				onSelectUser={(u: ChatCostUserRollup) => {
 					setSearchParams((prev) => {
 						const next = new URLSearchParams(prev);
 						next.set("user", u.user_id);
@@ -144,4 +229,4 @@ const AgentSettingsUsagePage: FC<AgentSettingsUsagePageProps> = ({ now }) => {
 	);
 };
 
-export default AgentSettingsUsagePage;
+export default AgentSettingsSpendPage;
