+++ site/src/pages/AgentsPage/components/ChatCostSummaryView.tsx
@@ -1,9 +1,10 @@
 import dayjs from "dayjs";
 import { TriangleAlertIcon } from "lucide-react";
-import type { FC } from "react";
+import { type FC, useState } from "react";
 import { getErrorMessage } from "#/api/errors";
 import type * as TypesGen from "#/api/typesGenerated";
 import { Button } from "#/components/Button/Button";
+import { PaginationWidgetBase } from "#/components/PaginationWidget/PaginationWidgetBase";
 import { Spinner } from "#/components/Spinner/Spinner";
 import {
 	Table,
@@ -13,6 +14,11 @@ import {
 	TableHeader,
 	TableRow,
 } from "#/components/Table/Table";
+import {
+	Tooltip,
+	TooltipContent,
+	TooltipTrigger,
+} from "#/components/Tooltip/Tooltip";
 import { formatTokenCount } from "#/utils/analytics";
 import { formatCostMicros } from "#/utils/currency";
 
@@ -52,6 +58,9 @@ export const ChatCostSummaryView: FC<ChatCostSummaryViewProps> = ({
 	loadingLabel,
 	emptyMessage,
 }) => {
+	const [modelPage, setModelPage] = useState(1);
+	const [chatPage, setChatPage] = useState(1);
+
 	if (isLoading) {
 		return (
 			<div
@@ -80,6 +89,17 @@ export const ChatCostSummaryView: FC<ChatCostSummaryViewProps> = ({
 	if (!summary) {
 		return null;
 	}
+
+	const modelPageSize = 10;
+	const pagedModels = summary.by_model.slice(
+		(modelPage - 1) * modelPageSize,
+		modelPage * modelPageSize,
+	);
+	const chatPageSize = 10;
+	const pagedChats = summary.by_chat.slice(
+		(chatPage - 1) * chatPageSize,
+		chatPage * chatPageSize,
+	);
 
 	const usageLimit = summary.usage_limit;
 	const showUsageLimitCard = usageLimit?.is_limited === true;
@@ -241,119 +261,131 @@ export const ChatCostSummaryView: FC<ChatCostSummaryViewProps> = ({
 				</p>
 			) : (
 				<>
-					<div className="overflow-x-auto rounded-lg border border-border-default">
-						<Table className="text-sm" aria-label="Cost breakdown by model">
+					<div>
+						<Table aria-label="Cost breakdown by model">
 							<TableHeader>
-								<TableRow className="text-left text-xs font-medium uppercase tracking-wide text-content-secondary">
-									<TableHead className="px-4 py-3">Model</TableHead>
-									<TableHead className="px-4 py-3">Provider</TableHead>
-									<TableHead className="px-4 py-3 text-right">Cost</TableHead>
-									<TableHead className="px-4 py-3 text-right">
-										Messages
-									</TableHead>
-									<TableHead className="px-4 py-3 text-right">Input</TableHead>
-									<TableHead className="px-4 py-3 text-right">Output</TableHead>
-									<TableHead className="px-4 py-3 text-right">
-										Cache Read
-									</TableHead>
-									<TableHead className="px-4 py-3 text-right">
-										Cache Write
-									</TableHead>
+								<TableRow>
+									<TableHead>Model</TableHead>
+									<TableHead>Provider</TableHead>
+									<TableHead className="text-right">Cost</TableHead>
+									<TableHead className="text-right">Messages</TableHead>
+									<TableHead className="text-right">Input</TableHead>
+									<TableHead className="text-right">Output</TableHead>
+									<TableHead className="text-right">Cache Read</TableHead>
+									<TableHead className="text-right">Cache Write</TableHead>
 								</TableRow>
 							</TableHeader>
 							<TableBody>
-								{summary.by_model.map((model) => (
-									<TableRow
-										key={model.model_config_id}
-										className="border-t border-border-default"
-									>
-										<TableCell className="px-4 py-3">
-											{model.display_name || model.model}
-										</TableCell>
-										<TableCell className="px-4 py-3 text-content-secondary">
+								{pagedModels.map((model) => (
+									<TableRow key={model.model_config_id}>
+										<TableCell>{model.display_name || model.model}</TableCell>
+										<TableCell className="text-content-secondary">
 											{model.provider}
 										</TableCell>
-										<TableCell className="px-4 py-3 text-right">
+										<TableCell className="text-right tabular-nums">
 											{formatCostMicros(model.total_cost_micros)}
 										</TableCell>
-										<TableCell className="px-4 py-3 text-right">
+										<TableCell className="text-right tabular-nums">
 											{model.message_count.toLocaleString()}
 										</TableCell>
-										<TableCell className="px-4 py-3 text-right">
+										<TableCell className="text-right tabular-nums">
 											{formatTokenCount(model.total_input_tokens)}
 										</TableCell>
-										<TableCell className="px-4 py-3 text-right">
+										<TableCell className="text-right tabular-nums">
 											{formatTokenCount(model.total_output_tokens)}
 										</TableCell>
-										<TableCell className="px-4 py-3 text-right">
+										<TableCell className="text-right tabular-nums">
 											{formatTokenCount(model.total_cache_read_tokens)}
 										</TableCell>
-										<TableCell className="px-4 py-3 text-right">
+										<TableCell className="text-right tabular-nums">
 											{formatTokenCount(model.total_cache_creation_tokens)}
 										</TableCell>
 									</TableRow>
 								))}
 							</TableBody>
 						</Table>
+						{summary.by_model.length > modelPageSize && (
+							<div className="pt-4">
+								<PaginationWidgetBase
+									totalRecords={summary.by_model.length}
+									currentPage={modelPage}
+									pageSize={modelPageSize}
+									onPageChange={setModelPage}
+									hasPreviousPage={modelPage > 1}
+									hasNextPage={
+										modelPage * modelPageSize < summary.by_model.length
+									}
+								/>
+							</div>
+						)}
 					</div>
 
-					<div className="overflow-x-auto rounded-lg border border-border-default">
-						<Table
-							className="text-sm"
-							aria-label="Cost breakdown by conversation"
-						>
+					<div>
+						<Table aria-label="Cost breakdown by conversation">
 							<TableHeader>
-								<TableRow className="text-left text-xs font-medium uppercase tracking-wide text-content-secondary">
-									<TableHead className="px-4 py-3">Conversation</TableHead>
-									<TableHead className="px-4 py-3 text-right">Cost</TableHead>
-									<TableHead className="px-4 py-3 text-right">
-										Messages
-									</TableHead>
-									<TableHead className="px-4 py-3 text-right">Input</TableHead>
-									<TableHead className="px-4 py-3 text-right">Output</TableHead>
-									<TableHead className="px-4 py-3 text-right">
-										Cache Read
-									</TableHead>
-									<TableHead className="px-4 py-3 text-right">
-										Cache Write
-									</TableHead>
+								<TableRow>
+									<TableHead>Conversation</TableHead>
+									<TableHead className="text-right">Cost</TableHead>
+									<TableHead className="text-right">Messages</TableHead>
+									<TableHead className="text-right">Input</TableHead>
+									<TableHead className="text-right">Output</TableHead>
+									<TableHead className="text-right">Cache Read</TableHead>
+									<TableHead className="text-right">Cache Write</TableHead>
 								</TableRow>
 							</TableHeader>
 							<TableBody>
-								{summary.by_chat.map((chat) => (
-									<TableRow
-										key={chat.root_chat_id}
-										className="border-t border-border-default"
-									>
-										<TableCell className="px-4 py-3">
-											{chat.chat_title || (
+								{pagedChats.map((chat) => (
+									<TableRow key={chat.root_chat_id}>
+										<TableCell className="max-w-[200px]">
+											{chat.chat_title ? (
+												<Tooltip>
+													<TooltipTrigger asChild>
+														<span className="block truncate">
+															{chat.chat_title}
+														</span>
+													</TooltipTrigger>
+													<TooltipContent>{chat.chat_title}</TooltipContent>
+												</Tooltip>
+											) : (
 												<span className="italic text-content-secondary">
 													Untitled conversation
 												</span>
 											)}
 										</TableCell>
-										<TableCell className="px-4 py-3 text-right">
+										<TableCell className="text-right tabular-nums">
 											{formatCostMicros(chat.total_cost_micros)}
 										</TableCell>
-										<TableCell className="px-4 py-3 text-right">
+										<TableCell className="text-right tabular-nums">
 											{chat.message_count.toLocaleString()}
 										</TableCell>
-										<TableCell className="px-4 py-3 text-right">
+										<TableCell className="text-right tabular-nums">
 											{formatTokenCount(chat.total_input_tokens)}
 										</TableCell>
-										<TableCell className="px-4 py-3 text-right">
+										<TableCell className="text-right tabular-nums">
 											{formatTokenCount(chat.total_output_tokens)}
 										</TableCell>
-										<TableCell className="px-4 py-3 text-right">
+										<TableCell className="text-right tabular-nums">
 											{formatTokenCount(chat.total_cache_read_tokens)}
 										</TableCell>
-										<TableCell className="px-4 py-3 text-right">
+										<TableCell className="text-right tabular-nums">
 											{formatTokenCount(chat.total_cache_creation_tokens)}
 										</TableCell>
 									</TableRow>
 								))}
 							</TableBody>
 						</Table>
+						{summary.by_chat.length > chatPageSize && (
+							<div className="pt-4">
+								<PaginationWidgetBase
+									totalRecords={summary.by_chat.length}
+									currentPage={chatPage}
+									pageSize={chatPageSize}
+									onPageChange={setChatPage}
+									hasPreviousPage={chatPage > 1}
+									hasNextPage={chatPage * chatPageSize < summary.by_chat.length}
+								/>
+							</div>
+						)}
 					</div>
 				</>
 			)}
